# .rifle/openapi.lock.yaml
openapi: 3.1.0
info:
  title: EIFD/TIF Increment Batch API
  version: 1.0.0
  description: |
    Starts and observes the nine-step EIFD/TIF increment batch. The service uses
    preloaded shared-live assessment, tax-code, agency-reference, frozen-valuation,
    and agency-equalized-valuation data. Requests never identify local files.
    Every launch pins the business clock for the complete chain. Runs share the
    live indexed stores and are not isolated snapshots.
servers:
- url: http://localhost:8080
  description: Local development
...
security: []
paths:
...
  /api/eifd-tif-increment-runs:
    post:
      tags:
      - EifdTifIncrementRuns
      operationId: startEifdTifIncrementRun
      summary: Start an EIFD/TIF increment batch run
      description: |
        Creates one asynchronous run over server-preloaded shared-live data. The
        pinned businessDate and businessTime apply to every step. The batch updates
        frozen valuations, calculates and reports tax-code and agency increments,
        and posts matched agency equalized valuations. A 201 response means that
        the run was accepted, not that processing completed. Read the returned
        resource with GET /api/eifd-tif-increment-runs/{id} to observe the final
        status, return code, record counts, logical outputs, and messages.

        Validation, ordered-merge behavior, arithmetic, partial-output behavior,
        omission rules, reporting, and shared-live persistence are governed by the
        accepted rules listed in x-business-rules.
      x-business-rules:
      - id: asrea740-001
        description: Require a complete twelve-character reassessment control whose two-character current year is numeric; reject it before assessment processing otherwise.
      - id: asrea740-002
        description: Require numeric ascending improvement-age and Class 288 year ranges; interpret two-digit years 61-99 as 1961-1999 and 00-60 as 2000-2060.
      - id: asrea740-003
        description: Report each division whose improvement components do not reconcile to its improvement total, without preventing frozen-value storage.
      - id: asrea740-004
        description: Classify reassessment-township improvement value by age, Class 288 status, improvement year, and permit status before forming frozen-value categories.
      - id: asrea740-005
        description: Truncate fractional prior-improvement amounts at every occupancy, condition, and assessment percentage stage; a missing positive factor leaves the preceding stage unchanged.
      - id: asrea740-006
        description: Select change-action totals for qualifying new value, current Class 288 activity, permit change, or the zero township prefix; otherwise select no-change totals, including the fully-exempt-to-nonexempt exception.
      - id: asrea740-007
        description: Add prior, current, proposed, change-action, and no-change division totals to an existing frozen valuation and replace the stored division.
      - id: asrea740-008
        description: Insert a missing frozen-valuation division only when at least one proposed or action value is nonzero.
      - id: asrea742-001
        description: Include every current-assessment property in the selection population and initialize uncalculated allocation amounts and percentages to zero.
      - id: asrea742-002
        description: Include current Class 288 assessed value only for category 3, class 288, the processing year, and material-permit-with-building properties.
      - id: asrea742-003
        description: Include first-year non-frozen assessed value only for categories 2-5, age one, and a class other than 288.
      - id: asrea742-004
        description: Exclude a paired questionnaire response from valuation calculations for a questionnaire-designated improvement.
      - id: asrea742-005
        description: Require processingYear to contain exactly two characters before selection begins.
      - id: asrea742-006
        description: Require the two-character processingYear to contain digits only before selection begins.
      - id: asrea743-001
        description: Require selections to be nondecreasing by division and tax code; exclude an offending lower key, fail the run, and retain eligible earlier and trailing partial products.
      - id: asrea743-002
        description: Accumulate consecutive duplicate division-and-tax-code keys into one group without increasing the division tax-code group count.
      - id: asrea743-003
        description: Sum current-288 and first-time assessed values for each division-and-tax-code group, excluding expired-288 value from both totals.
      - id: asrea743-004
        description: On a tax-code change within a division, emit only the prior tax-code rollup and continue the division rollup with its group count incremented.
      - id: asrea743-005
        description: On a division change, emit the prior tax-code and division rollups, including the division count and last tax code, then start both new groups.
      - id: asrea743-006
        description: At end of input or ordering failure, independently emit each trailing tax-code and division rollup only when its first-time total is positive.
      - id: asrea743-007
        description: Suppress a trailing rollup whose first-time total is zero even when current-288 value is nonzero; empty input emits no rollups.
      - id: asrea744-001
        description: Require both total streams to be strictly ascending by division and tax code; duplicate or descending keys fail before later percentages or completion totals.
      - id: asrea744-002
        description: Pair each tax-code total with the division total having the same division, preserving tax-code identity and count and initializing later monetary allocations to zero.
      - id: asrea744-003
        description: Calculate current-288 and first-time shares only when the matching division total is positive; otherwise use zero without division.
      - id: asrea744-004
        description: Truncate current-288 and first-time shares toward zero at seven fractional decimal places.
      - id: asrea744-005
        description: Advance a lower division total; report and omit a lower unmatched tax-code total; emit a percentage row only for equal divisions.
      - id: asrea745-001
        description: Give a sole tax code the division's full proposed first-time, current-288, and expired-288 values, with expired incentive initialized to zero.
      - id: asrea745-002
        description: For multiple tax codes, multiply first-time and current-288 values by each positive percentage; a nonpositive percentage yields zero with no even-share fallback.
      - id: asrea745-003
        description: For multiple tax codes, divide expired-288 value evenly by tax-code count and discard the remainder, while initializing expired incentive to zero.
      - id: asrea745-004
        description: Silently omit an allocation whose division lacks a valuation record, continue processing, and expose the omission only in the aggregate unmatched count.
      - id: asrea748-001
        description: Require annualEqualizationFactor to be exactly five numeric characters and greater than zero before reading any frozen-agency input.
      - id: asrea748-002
        description: Multiply grouped current-288, expired-288, and first-time frozen assessed value by the annual factor and round half away from zero to whole dollars.
      - id: asrea748-003
        description: Calculate frozen tax from frozen assessed value times the tax-code rate divided by 100, rounding half away from zero to two decimals; do not use equalized value as the base.
      - id: asrea748-004
        description: Examine all 40 tax-code agency positions in order and emit one frozen-agency record for each positive agency number, skipping zero and negative entries.
      - id: asrea748-005
        description: For a missing tax-code master, report the tax code, retain earlier records, emit none for that tax code, skip later tax codes, and fail the run.
      - id: clrtm749-001
        description: Require agency-detail input to be nondecreasing by agency; a lower agency stops processing and prevents publication of later groups.
      - id: clrtm749-002
        description: Sum all nine assessed and equalized increment categories once for each consecutive agency group without rounding.
      - id: clrtm749-003
        description: Require an agency-reference description before publishing a summarized agency; a missing reference fails instead of publishing or skipping the group.
      - id: clrtm749-004
        description: Keep new-property and expired-incentive equalized values separate in agency summaries so the posting step can combine them.
      - id: clrtm749-005
        description: Produce one agency-keyed machine-readable summary for each completed referenced agency group and pass it to agency valuation posting.
      - id: clrtm750-001
        description: Require agency-town report input to be nondecreasing by agency and town; a lower key ends reporting with an error.
      - id: clrtm750-002
        description: Sum five equalized-value categories per agency-town group and sum town totals into each agency total without rounding.
      - id: clrtm750-003
        description: Resolve suburban town codes 10-39 at town minus 9 and city codes 70-77 at town minus 39 in the shared name table; values outside those ranges are not separately rejected before applying the city offset.
      - id: clrtm750-004
        description: Emit a paged agency-and-town report and agency-year append records containing headings, town details, and totals for all five categories.
      - id: clrtm756-001
        description: Require posting inputs to be strictly increasing by agency; a duplicate or lower agency stops before its values are applied.
      - id: clrtm756-002
        description: Replace persistent new-property EAV with the sum of input new-property EAV and expired-incentive EAV without rounding.
      - id: clrtm756-003
        description: Carry input annexed-property EAV, disconnected-property EAV, and disconnected TIF difference directly into their persistent fields.
      - id: clrtm756-004
        description: Replace the four posting fields on each matched persistent agency record while retaining all other stored values.
      - id: clrtm756-005
        description: Report and skip an input agency with no persistent valuation record, insert nothing, and continue with the next agency.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EifdTifIncrementRunRequest'
            examples:
              capturedTrace:
                summary: Pinned controls corresponding to the captured successful trace
                value:
                  businessDate: 2025-09-15
                  businessTime: 12:00:00
                  idempotencyKey: eifd-2025-09-15-noon
                  reassessmentControl: '260181202526'
                  processingYear: '26'
                  reportingYear: '2026'
                  annualEqualizationFactor: '10000'
      responses:
        '201':
          description: Run created and queued or started. Observe its final outcome through the returned resource URL.
          headers:
            Location:
              description: URL of the asynchronous run resource.
              required: true
              schema:
                type: string
              example: /api/eifd-tif-increment-runs/9000000000000001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EifdTifIncrementRunResponse'
              example:
                id: 9000000000000001
                status: QUEUED
                businessDate: 2025-09-15
                businessTime: 12:00:00
                outputs: []
                messages: []
        '400':
          description: A launch control violates a business rule, so no run is created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EifdTifIncrementErrorResponse'
              examples:
                invalidReassessmentControl:
                  value:
                    error: INVALID_REASSESSMENT_CONTROL
                    message: reassessmentControl must be a complete twelve-character value with a numeric current year.
                    ruleId: asrea740-001
                invalidReassessmentRanges:
                  value:
                    error: INVALID_REASSESSMENT_RANGES
                    message: Improvement-age and Class 288 year bounds must be numeric and ascending.
                    ruleId: asrea740-002
                invalidProcessingYear:
                  value:
                    error: INVALID_PROCESSING_YEAR
                    message: processingYear must contain exactly two characters.
                    ruleId: asrea742-005
                nonnumericProcessingYear:
                  value:
                    error: INVALID_PROCESSING_YEAR
                    message: processingYear must contain digits only.
                    ruleId: asrea742-006
                invalidEqualizationFactor:
                  value:
                    error: INVALID_EQUALIZATION_FACTOR
                    message: annualEqualizationFactor must be five numeric characters and greater than zero.
                    ruleId: asrea748-001
        '409':
          description: The idempotency key is already bound to a different EIFD/TIF increment launch request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EifdTifIncrementErrorResponse'
              example:
                error: IDEMPOTENCY_CONFLICT
                message: idempotencyKey is already associated with different launch controls.
  /api/eifd-tif-increment-runs/{id}:
    get:
      tags:
      - EifdTifIncrementRuns
      operationId: getEifdTifIncrementRun
      summary: Observe an EIFD/TIF increment batch run
      description: |
        Returns the current state of the asynchronous run. This operation only
        observes execution; all EIFD business rules apply to the POST launch.
        COMPLETED and FAILED are terminal. A terminal response carries the final
        return code, aggregate record activity, logical output metadata, and
        rule-linked messages. Earlier shared-live updates and outputs can remain
        visible after a later step fails when the applicable POST rule specifies
        partial products or continued processing.
      parameters:
      - name: id
        in: path
        required: true
        description: Positive 64-bit surrogate identifier returned by the POST operation.
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: Current run status and, when terminal, its final observed result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EifdTifIncrementRunResponse'
              examples:
                completedCapturedTrace:
                  summary: Successful observation consistent with the captured nine-step trace
                  value:
                    id: 9000000000000001
                    status: COMPLETED
                    businessDate: 2025-09-15
                    businessTime: 12:00:00
                    returnCode: 0
                    outputs:
                    - name: frozenValuationReconciliationReport
                      recordCount: 239
                    - name: agencySummaryReport
                      recordCount: 14
                    - name: townWithinAgencyReport
                      recordCount: 75
                    - name: agencyYearAppendRecords
                      recordCount: 117
                    messages:
                    - severity: INFO
                      text: EIFD/TIF increment batch completed successfully.
        '404':
          description: No EIFD/TIF increment run exists for the supplied surrogate identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EifdTifIncrementErrorResponse'
              example:
                error: RUN_NOT_FOUND
                message: No EIFD/TIF increment run was found for id 9000000000000001.
components:
  schemas:
...
    EifdTifIncrementRunRequest:
      type: object
      additionalProperties: false
      required:
      - businessDate
      - businessTime
      - idempotencyKey
      - reassessmentControl
      - processingYear
      - reportingYear
      - annualEqualizationFactor
      properties:
        businessDate:
          type: string
          format: date
          description: Pinned business date applied to every program in the run.
          example: 2025-09-15
        businessTime:
          type: string
          pattern: ^([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
          description: Pinned 24-hour business time applied to every program in the run.
          example: 12:00:00
        idempotencyKey:
          type: string
          minLength: 1
          description: Caller-selected key for deduplicating equivalent launch attempts.
        reassessmentControl:
          type: string
          minLength: 12
          maxLength: 12
          pattern: ^[0-9]{12}$
          description: |
            Complete ASREA740 reassessment control containing the current year,
            improvement-age range, and Class 288 year range. The service enforces
            ascending ranges and the century pivot in asrea740-002.
          example: '260181202526'
        processingYear:
          type: string
          minLength: 2
          maxLength: 2
          pattern: ^[0-9]{2}$
          description: Two-digit ASREA742 processing year used for Class 288 selection.
          example: '26'
        reportingYear:
          type: string
          minLength: 4
          maxLength: 4
          pattern: ^[0-9]{4}$
          description: Four-digit reporting year used by the agency summary and town-within-agency report.
          example: '2026'
        annualEqualizationFactor:
          type: string
          minLength: 5
          maxLength: 5
          pattern: ^[0-9]{5}$
          description: Five-character positive factor with one integer and four fractional digits; 00000 is rejected by asrea748-001.
          example: '10000'
    EifdTifIncrementRunResponse:
      type: object
      additionalProperties: false
      required:
      - id
      - status
      - businessDate
      - businessTime
      - outputs
      - messages
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Positive 64-bit surrogate run identifier.
          example: 9000000000000001
        status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
        businessDate:
          type: string
          format: date
          description: Business date pinned by the launch request.
        businessTime:
          type: string
          pattern: ^([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
          description: Business time pinned by the launch request.
        returnCode:
          type: integer
          minimum: 0
          description: Final maximum batch return code; omitted before the run is terminal.
        recordsRead:
          type: integer
          minimum: 0
          description: Aggregate records read across completed steps; omitted until available.
        recordsWritten:
          type: integer
          minimum: 0
          description: Aggregate records written across completed steps; omitted until available.
        recordsUpdated:
          type: integer
          minimum: 0
          description: Aggregate shared-live records updated across completed steps; omitted until available.
        recordsRejected:
          type: integer
          minimum: 0
          description: Aggregate records rejected or omitted by a reported rule path; omitted until available.
        outputs:
          type: array
          description: Logical batch products and reports; items never expose server filesystem paths.
          items:
            $ref: '#/components/schemas/EifdTifIncrementOutput'
        messages:
          type: array
          description: Operational and business-rule messages observed so far.
          items:
            $ref: '#/components/schemas/EifdTifIncrementMessage'
    EifdTifIncrementOutput:
      type: object
      additionalProperties: false
      required:
      - name
      - recordCount
      properties:
        name:
          type: string
          description: Stable logical product name, such as agencySummaryReport; never a local path.
        recordCount:
          type: integer
          minimum: 0
        generation:
          type: integer
          minimum: 1
          description: Shared-live logical generation when the product is versioned.
    EifdTifIncrementMessage:
      type: object
      additionalProperties: false
      required:
      - severity
      - text
      properties:
        severity:
          type: string
          enum:
          - INFO
          - WARNING
          - ERROR
        text:
          type: string
        ruleId:
          type: string
          description: Exact accepted EIFD business-rule identifier when the message reports a rule outcome.
          example: asrea740-003
    EifdTifIncrementErrorResponse:
      type: object
      additionalProperties: false
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
        ruleId:
          type: string
          description: Exact accepted EIFD rule when a business-rule violation caused the error; omitted for operational errors.
          enum:
          - asrea740-001
          - asrea740-002
          - asrea742-005
          - asrea742-006
          - asrea748-001
...